testing/model_tests/test_pipeline/requirements.txt
pinecone-client>=3.0.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
unstructured>=0.10.0
pypdf>=3.0.0 